Update to Bevy 0.10 (#77)
* Update to Bevy 0.10

* Fix CI and apply clippy / formatting

`rust-toolchain` action does not run any commands, it just installs
the toolchain. Also, it should run before caching crates.

* Remove profile settings

It has no effect unless it is in the root of the workspace.
Remove it to remove the warning about it. The behavior hasn't changed

+++ renetcode/src/packet.rs
@@ -344,7 +344,7 @@ fn get_additional_data(prefix: u8, protocol_id: u64) -> [u8; 13 + 8 + 1] {
 }
 
 fn decode_prefix(value: u8) -> (u8, usize) {
-    ((value & 0xF) as u8, (value >> 4) as usize)
+    ((value & 0xF), (value >> 4) as usize)
 }
 
 fn encode_prefix(value: u8, sequence: u64) -> u8 {
